Links sparkjava to build so that it is published
sparkjava wasn't added to the root pom, so it never built, nor was
published.

This also uses standard formatting and fixes the naming convention on
the integration test so it isn't run twice.
