+++ brave-sparkjava/pom.xml
@@ -5,7 +5,7 @@
   <parent>
     <artifactId>brave-parent</artifactId>
     <groupId>io.zipkin.brave</groupId>
-    <version>4.0.7-SNAPSHOT</version>
+    <version>4.1.1-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
