+++ brave-sparkjava/src/main/java/com/github/kristofa/brave/sparkjava/BraveTracing.java
