+++ pom.xml
@@ -110,6 +110,7 @@
     <module>brave-okhttp</module>
     <module>brave-p6spy</module>
     <module>brave-cxf3</module>
+    <module>brave-sparkjava</module>
   </modules>
 
   <distributionManagement>
